fixed build and resolved pr issues (could not handle redis config IT test, sorry ;))
